Cleanup, seperate black/white lists setup
